Cleanup: move Toast components to common/Toast
